refactor: Move logic of CsvExporter to ExportDialog
