+++ src/components/ExportCSV/ExportDialog.jsx
@@ -1,12 +1,41 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 
 import Buttons from 'cozy-ui/transpiled/react/Buttons'
 import Typography from 'cozy-ui/transpiled/react/Typography'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import { ConfirmDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
+import { generateWebLink, useClient } from 'cozy-client'
+import { exportTripsToCSV } from 'src/lib/exportTripsToCSV'
 
-const ExportDialog = ({ onClose, appDir, fileCreated, linkToAppFolder }) => {
+const ExportDialog = ({ onClose, accountName }) => {
   const { t } = useI18n()
+  const client = useClient()
+  const [linkToAppFolder, setLinkToAppFolder] = useState('')
+  const [{ file, appDir }, setFileCreated] = useState({
+    file: null,
+    appDir: null
+  })
+
+  useEffect(() => {
+    if (!linkToAppFolder && appDir) {
+      const link = generateWebLink({
+        slug: 'drive',
+        cozyUrl: client.getStackClient().uri,
+        subDomainType: client.getInstanceOptions().subdomain,
+        pathname: '/',
+        hash: `folder/${appDir._id}`
+      })
+      setLinkToAppFolder(link)
+    }
+  }, [appDir, client, linkToAppFolder])
+
+  useEffect(() => {
+    const exportTrips = async () => {
+      const { appFolder, file } = await exportTripsToCSV(client, t, accountName)
+      setFileCreated({ file, appDir: appFolder })
+    }
+    exportTrips()
+  }, [accountName, client, t])
 
   return (
     <ConfirmDialog
@@ -17,7 +46,7 @@ const ExportDialog = ({ onClose, appDir, fileCreated, linkToAppFolder }) => {
         appDir ? (
           <Typography color="textPrimary">
             {t('export.modal.content.done', {
-              filename: fileCreated.name,
+              filename: file.name,
               pathAppDir: appDir.path
             })}
           </Typography>
